+++ packages/sdk/CHANGELOG.md
@@ -6,6 +6,7 @@ Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how 
 
 ## Recent Changes
 
+- Add new functions to ZsshUtils called `detectServerOnPath()` and `hasWriteAccess()`.
 - Added support to list APF and PROCLIB data sets. [#1029](https://github.com/zowe/zowex/pull/1029)
 - Added support for viewing syslog.
 - Added response content for copy data sets and members operations. [#995](https://github.com/zowe/zowex/pull/995)